requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
pyyaml==5.4.1
openai==0.27.0
pillow==8.3.1
pytesseract==0.3.8
selenium==4.1.0
beautifulsoup4==4.9.3
pydantic==1.8.2
sqlalchemy==1.4.23
pytest==6.2.5